+++ CHANGELOG.md
@@ -3,6 +3,8 @@
  * Upgraded to Gulp 4.
  * Update dependencies
  * Fixed lint warnings.
+ * Updated filesystem watching to use new `appcd.fs.watch()` and `appcd.fs.unwatch()` to optimize
+   subscriptions. [(DAEMON-253)](https://jira.appcelerator.org/browse/DAEMON-253)
 
 # v1.2.0 (Oct 25, 2018)